backend-core/pyproject.toml
[project]
name = "backend-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.41.2",
    "celery>=5.5.3",
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "redis>=7.1.0",
]
